+++ plugins/Makefile.am
@@ -33,7 +33,8 @@ MINGW_CFLAGS = \
 			-I$(top_srcdir)/tagmanager/src \
 			-I$(top_srcdir)/scintilla/include \
 			$(GTK_CFLAGS) \
-			$(PLUGIN_CFLAGS)
+			$(PLUGIN_CFLAGS) \
+			-fvisibility=default
 
 .PHONY: all clean
 
@@ -83,13 +84,13 @@ saveactions_la_CFLAGS   = -DG_LOG_DOMAIN=\""SaveActions"\"
 filebrowser_la_CFLAGS   = -DG_LOG_DOMAIN=\""FileBrowser"\"
 splitwindow_la_CFLAGS   = -DG_LOG_DOMAIN=\""SplitWindow"\"
 
-demoplugin_la_LIBADD    = $(GTK_LIBS)
-classbuilder_la_LIBADD  = $(GTK_LIBS)
-htmlchars_la_LIBADD     = $(GTK_LIBS)
-export_la_LIBADD        = $(GTK_LIBS) -lm
-saveactions_la_LIBADD   = $(GTK_LIBS)
-filebrowser_la_LIBADD   = $(GTK_LIBS)
-splitwindow_la_LIBADD   = $(GTK_LIBS)
+demoplugin_la_LIBADD    = $(top_builddir)/src/libgeany.la $(GTK_LIBS)
+classbuilder_la_LIBADD  = $(top_builddir)/src/libgeany.la $(GTK_LIBS)
+htmlchars_la_LIBADD     = $(top_builddir)/src/libgeany.la $(GTK_LIBS)
+export_la_LIBADD        = $(top_builddir)/src/libgeany.la $(GTK_LIBS) -lm
+saveactions_la_LIBADD   = $(top_builddir)/src/libgeany.la $(GTK_LIBS)
+filebrowser_la_LIBADD   = $(top_builddir)/src/libgeany.la $(GTK_LIBS)
+splitwindow_la_LIBADD   = $(top_builddir)/src/libgeany.la $(GTK_LIBS)
 
 endif # MINGW
 endif # PLUGINS
