+++ tagmanager/mio/Makefile.am
@@ -1,9 +1,9 @@
-noinst_LIBRARIES = libmio.a
+noinst_LTLIBRARIES = libmio.la
 
 AM_CPPFLAGS = -DG_LOG_DOMAIN=\"MIO\" #-DMIO_DEBUG
-AM_CFLAGS   = $(GTK_CFLAGS)
+AM_CFLAGS   = $(GTK_CFLAGS) -fvisibility=hidden
 
-libmio_a_SOURCES = mio.c
+libmio_la_SOURCES = mio.c
 
 EXTRA_DIST = \
 	mio.h \
